psanitize fixes

+++ src/lib/plugins/BasicJobTypes/GaussianJobType.cpp
@@ -131,32 +131,35 @@ ERetStatus CGaussianJobType::DetectJobType(CJob& job,bool& detected,std::ostream
 
     CFileName job_file = arg_job.GetFileNameWithoutExt();
 
-    ofstream ofs(job_file);
-    if( ! ofs ) {
-        sout << endl;
-        sout << "<b><red> ERROR: The gaussian job was detected but unable to open the job script '" << job_file << "' for writing!</red></b>" << endl;
-        sout << "<b><red>        Remove the file and check if you have the write permission in the job directory.</red></b>" << endl;
-        return(ERS_FAILED);
-    }
+    // do not write the file if psanitize is used
+    if( job.GetItem("basic/arguments","INF_SANITIZE_JOB","YES") != "YES" ){
+        ofstream ofs(job_file);
+        if( ! ofs ) {
+            sout << endl;
+            sout << "<b><red> ERROR: The gaussian job was detected but unable to open the job script '" << job_file << "' for writing!</red></b>" << endl;
+            sout << "<b><red>        Remove the file and check if you have the write permission in the job directory.</red></b>" << endl;
+            return(ERS_FAILED);
+        }
 
-    ofs << "#!/usr/bin/env infinity-env" << endl;
-    ofs << "# ----------------------------------------------------------" << endl;
-    ofs << endl;
-    ofs << "# activate gaussian module -----------------------" << endl;
-    ofs << "module add " << gmodule << ":" << gmodver << endl;
-    ofs << endl;
-    ofs << "# start job --------------------------------------" << endl;
-    ofs << gexec << " " << job_file << endl;
-    ofs << endl;
-    ofs << "# clean ------------------------------------------" << endl;
-    ofs << "rm -f core" << endl;
-    ofs << endl;
+        ofs << "#!/usr/bin/env infinity-env" << endl;
+        ofs << "# ----------------------------------------------------------" << endl;
+        ofs << endl;
+        ofs << "# activate gaussian module -----------------------" << endl;
+        ofs << "module add " << gmodule << ":" << gmodver << endl;
+        ofs << endl;
+        ofs << "# start job --------------------------------------" << endl;
+        ofs << gexec << " " << job_file << endl;
+        ofs << endl;
+        ofs << "# clean ------------------------------------------" << endl;
+        ofs << "rm -f core" << endl;
+        ofs << endl;
 
-    if( ! ofs ) {
-        sout << endl;
-        sout << "<b><red> ERROR: The gaussian job was detected but unable to write data to the job script '" << job_file << "'!</red></b>" << endl;
-        sout << "<b><red>        Check if you have the write permission to the job input directory.</red></b>" << endl;
-        return(ERS_FAILED);
+        if( ! ofs ) {
+            sout << endl;
+            sout << "<b><red> ERROR: The gaussian job was detected but unable to write data to the job script '" << job_file << "'!</red></b>" << endl;
+            sout << "<b><red>        Check if you have the write permission to the job input directory.</red></b>" << endl;
+            return(ERS_FAILED);
+        }
     }
 
     job.SetItem("basic/jobinput","INF_JOB_NAME",job_file);
@@ -205,27 +208,30 @@ bool CGaussianJobType::CheckInputFile(CJob& job,std::ostream& sout)
         }
     }
 
+    // what amount should be dedicated to the gaussian job
+    int perc = 95; // in %
+
     // check memory keyword
     CSmallString smem = job.GetItem("specific/resources","INF_MEMORY");
     long int mem = CResourceValue::GetSize(smem)*1024;
     long int umem = GetMemory(job,job_name);
 
-    if( abs(umem/1024/1024 - mem*95/100/1024/1024) > 2 ){
+    if( abs(umem/1024/1024 - mem*perc/100/1024/1024) > 2 ){
         sout << endl;
         sout << "<b><blue> WARNING: Inconsistency in the amount of requested memory was detected</blue></b>" << endl;
         sout << "<b><blue>          in the gaussian input file!</blue></b>" << endl;
         sout << endl;
-        sout << "<b><blue>          The ammount of memory requested via psubmit command (95%)  : " << setw(7) << mem*90/100/1024/1024 << " MB</blue></b>" << endl;
+        sout << "<b><blue>          The ammount of memory requested via psubmit command (95%)  : " << setw(7) << mem*perc/100/1024/1024 << " MB</blue></b>" << endl;
         sout << "<b><blue>          The ammount of memory requested in the gaussian input file : " << setw(7) << umem/1024/1024 << " MB (via %Mem)</blue></b>" << endl;
 
-        if( UpdateMemory(job,job_name,mem*90/100) == false ){
+        if( UpdateMemory(job,job_name,mem*perc/100) == false ){
             sout << endl;
             sout << "<b><red> ERROR: Unable to save updated gaussian input file (%Mem)!</red></b>" << endl;
             return(false);
         }
     }
 
-    if( ( abs(umem/1024/1024 - mem*90/100/1024/1024) > 2 ) || (uncpus != ncpus) ){
+    if( ( abs(umem/1024/1024 - mem*perc/100/1024/1024) > 2 ) || (uncpus != ncpus) ){
         sout << endl;
         sout << "<b><blue> WARNING: The input file was updated!</blue></b>" << endl;
     }
